Update PCBs to purchased revision
Main edit is using a ULN2003 NPN transistor array over discrete MOSFETs.
This was chosen because there was limited board space to put in
pull-down resistors on the board and it was complicating the single
layer routing. The PCB had to be expanded to 70x75mm anyways to
accommodate the SOIC ic anyways.

The schematics of the IO board were also updated to use net labels over
direct wiring. Due to the JST connectors and the Pi Pico schematic
symbol being mirrored, it was getting too difficult to have an
understandable schematic.

Minor edit was changing LED_DATA to nLED_DATA to signify that it is
an active low signal.